# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "getting-started",
    "description": "A playground for getting started with Zenable",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/findings": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "List Findings",
        "description": "List all guardrail findings.\n\nArgs:\n    request: FastAPI request.\n    api_key: Validated API key.\n    limit: Maximum findings to return.\n\nReturns:\n    List of findings.",
        "operationId": "list_findings_api_v1_findings_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingResponse"
                  },
                  "title": "Response List Findings Api V1 Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/findings/{finding_id}/decision": {
      "post": {
        "tags": [
          "findings"
        ],
        "summary": "Record Decision",
        "description": "Record a decision on a finding (approve/reject).\n\nArgs:\n    request: FastAPI request.\n    finding_id: The finding ID.\n    decision: The decision details.\n    api_key: Validated API key.\n\nReturns:\n    Confirmation message.",
        "operationId": "record_decision_api_v1_findings__finding_id__decision_post",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Record Decision Api V1 Findings  Finding Id  Decision Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kv": {
      "get": {
        "tags": [
          "kv"
        ],
        "summary": "List Kv Store",
        "description": "List all key-value pairs.\n\nArgs:\n    request: FastAPI request.\n    api_key: Validated API key.\n\nReturns:\n    List of KV entries.",
        "operationId": "list_kv_store_api_v1_kv_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KVStoreEntry"
                  },
                  "title": "Response List Kv Store Api V1 Kv Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kv/{key}": {
      "get": {
        "tags": [
          "kv"
        ],
        "summary": "Get Kv Value",
        "description": "Get a value by key.\n\nArgs:\n    request: FastAPI request.\n    key: The key to retrieve.\n    api_key: Validated API key.\n\nReturns:\n    The KV entry.\n\nRaises:\n    HTTPException: If key not found.",
        "operationId": "get_kv_value_api_v1_kv__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KVStoreEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "kv"
        ],
        "summary": "Set Kv Value",
        "description": "Set a key-value pair.\n\nArgs:\n    request: FastAPI request.\n    key: The key to set.\n    item: The value to store.\n    api_key: Validated API key.\n\nReturns:\n    Confirmation message.",
        "operationId": "set_kv_value_api_v1_kv__key__post",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KVStoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Set Kv Value Api V1 Kv  Key  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "kv"
        ],
        "summary": "Delete Kv Value",
        "description": "Delete a key-value pair.\n\nArgs:\n    request: FastAPI request.\n    key: The key to delete.\n    api_key: Validated API key.\n\nReturns:\n    Confirmation message.\n\nRaises:\n    HTTPException: If key not found.",
        "operationId": "delete_kv_value_api_v1_kv__key__delete",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Kv Value Api V1 Kv  Key  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Trigger Scan",
        "description": "Trigger a new scan.\n\nArgs:\n    request: FastAPI request.\n    scan_request: Scan parameters.\n    background_tasks: FastAPI background tasks.\n    api_key: Validated API key.\n\nReturns:\n    Scan job details.",
        "operationId": "trigger_scan_api_v1_scans_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Scan Api V1 Scans Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{job_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan Status",
        "description": "Get the status of a scan job.\n\nArgs:\n    request: FastAPI request.\n    job_id: The batch job ID.\n    api_key: Validated API key.\n\nReturns:\n    Job status and results.\n\nRaises:\n    HTTPException: If job not found.",
        "operationId": "get_scan_status_api_v1_scans__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Scan Status Api V1 Scans  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FindingDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision",
            "description": "Decision: 'approve' or 'reject'"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Optional decision note"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "FindingDecisionRequest",
        "description": "Request to record a finding decision."
      },
      "FindingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "line_number": {
            "type": "integer",
            "title": "Line Number"
          },
          "pattern_name": {
            "type": "string",
            "title": "Pattern Name"
          },
          "line_content": {
            "type": "string",
            "title": "Line Content"
          },
          "scanned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scanned At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_path",
          "line_number",
          "pattern_name",
          "line_content",
          "scanned_at"
        ],
        "title": "FindingResponse",
        "description": "A guardrail finding."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KVStoreEntry": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "updated_at"
        ],
        "title": "KVStoreEntry",
        "description": "A key-value store entry."
      },
      "KVStoreRequest": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value",
            "description": "The value to store"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "KVStoreRequest",
        "description": "Request to set/update a KV entry."
      },
      "ScanRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Directory to scan",
            "default": "."
          }
        },
        "type": "object",
        "title": "ScanRequest",
        "description": "Request to start a new scan."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}